djsongrequest.py
Create a program that manages song requests for a DJ. The program should:

Continuously ask the user to enter a song title. Exit the program if the user enters Quit.
If the song title was previously requested, notify the user and share the song position.

Example: “The song Hero has already been requested. It is song number 3 on the list.”
Otherwise, add the song to the request list, notify the user and share the song position.

Example: “The song Fantasy has been added. It is song number 10 on the list.”

Directions -

- Begin planning out your program. Consider how you might store and access the list. You may also want to include your plan in code comments in the repl.
- Start coding your song request program. You may need to create multiple methods in your program!
- Test your program by running calls to your function and analyzing the output. Be sure to test at least 3 cases