is_anagram.py
Write a function called is_anagram() that takes two string arguments, str1 and str2. The function check whether str1 and str2 are an anagram of each other or not. If the strings are anagrams, return True. Otherwise, False.

An anagram of a string is another string that contains the same characters, only the order of characters can be different.