diagonal_sum.py
Write a function called diagonal_sum() that takes in a 2D list m as a parameter. The function returns the sum of the elements on the left diagonal of m. You can assume that the given 2D list has the same number of rows and columns.

Input: [ [1, 2, 3],
         [4, 5, 6],
         [7, 8, 9]]
Output: 15
**Explanation**: Elements at m[0][0], m[1][1], and m[2][2] are along the
left diagonal of me. Thus, 1 + 5 + 9 is equal to 15.

Input: [ [1] ]
Output: 1
**Element**: Element at m[0][0] is along the left diagonal of m. Thus, 1 is returned.