rotate90.py
Write a function called rotate_matrix() that takes a 2D list m as a parameter. The function returns the modified version of the list where elements are turned by 90 degrees in the counter-clockwise direction. You can assume that the given 2D list has the same number of rows and columns.