connect3.py
Given a list 'moves', in which there are a series of moves, represented as lists, return True if there are 3 moves that are adjacent to each other. This means 3 moves that are in a row, column, or diagonal. Return False if there are no moves that are 3 in a row.

Input: [[0,0], [0,2], [1,1], [1,2], [2,2]]
Output: True
Explanation: There are two options that show three adjacent moves.
The moves (0,0), (1,1), (2,2) are adjacent in the diagonal
The moves (0,2), (1, 2), (2,2) are adjacent in the column

Input: [[0,0], [0,2], [1,0], [2,0]]
Output: True
Explanation: The moves (0,0), (1,0), and (2,0) are adjacent in the column

Input: [[0,0], [0,1], [1,1], [0,2]] 
Output: False
Explanation: There are no moves that show three adjacent moves.